Implemented two functions inputCheck & boundsCheck
The inputCheck function: ensures that a user may only enter an input of
type int. If this is done successfully then the user is passed to the
boundsCheck function.

The boundsCheck function: ensures that user enters an integer between 0
and getCols(). The int cols variable is the max size of the user input
otherwise the Triangular Matrix would be no longer viable. If the user
enters an input to meet the set criteria then, int rows variable is set
to the user’s input. If unsuccessful the user is sent back to the
inputCheck function to repeat the process.