Added some interaction
Added some new functionality into the main class so that the program
would have a bit more usage to it(Options menu for user interaction and
dynamic selection of what type of vector for template class to use and
what size to create ). Also changed up some things in the .h file
(added in a menu display and menu control).